refactor: padding to crypto.cpp file
