+++ CMakeLists.txt
@@ -23,10 +23,6 @@ target_include_directories(libio PRIVATE ${Boost_INCLUDE_DIR})
 target_link_libraries(libio PRIVATE ${Boost_LIBRARIES})
 # set_target_properties(libio PROPERTIES LINKER_LANGUAGE CXX)
 
-# padding
-add_library(libpadding ${CMAKE_SOURCE_DIR}/lib/crypto/padding.hpp)
-set_target_properties(libpadding PROPERTIES LINKER_LANGUAGE CXX)
-
 # ciphermode
 add_library(libciphermode ${CMAKE_SOURCE_DIR}/lib/crypto/ciphermode.hpp
                           ${CMAKE_SOURCE_DIR}/lib/crypto/ciphermode.cpp)
@@ -36,10 +32,6 @@ add_library(libciphermode ${CMAKE_SOURCE_DIR}/lib/crypto/ciphermode.hpp
 # io
 add_executable(test_io ${CMAKE_SOURCE_DIR}/lib/io/test_io.cpp)
 target_link_libraries(test_io PRIVATE Catch2::Catch2WithMain libio)
-
-# padding
-add_executable(test_padding ${CMAKE_SOURCE_DIR}/lib/crypto/test_padding.cpp)
-target_link_libraries(test_padding PRIVATE Catch2::Catch2WithMain libpadding)
 
 # crypto
 add_executable(test_crypto ${CMAKE_SOURCE_DIR}/lib/crypto/test_crypto.cpp
@@ -56,4 +48,4 @@ target_link_libraries(test_ciphermode PRIVATE Catch2::Catch2WithMain
 
 set(bin aes-cli)
 add_executable(${bin} ${CMAKE_SOURCE_DIR}/app/main.cpp)
-target_link_libraries(${bin} libio libpadding libciphermode)
+target_link_libraries(${bin} libio libciphermode)
